+++ api/V0ScaleEtcd.go
@@ -30,13 +30,12 @@ func V0ScaleEtcd(w http.ResponseWriter, r *http.Request) {
 
 		d = []byte(strconv.Itoa(newCount - oldCount))
 
-		err := SaveConfig()
-		if err != nil {
-			d = ErrorMessage(1, "V0ScaleEtcd", "Could not save config data")
-		}
+		// Save current config
+		SaveConfig()
+
 		// if scale down, kill not needes agents
 		if newCount < oldCount {
-			keys := GetAllRedisKeys(config.PrefixHostname + "agent:*")
+			keys := GetAllRedisKeys(framework.FrameworkName + ":etcd:*")
 
 			for keys.Next(config.RedisCTX) {
 				if newCount < oldCount {