+++ app.go
@@ -2,12 +2,16 @@ package main
 
 import (
 	"context"
+	"crypto/tls"
+	"encoding/base64"
 	"encoding/json"
+	"flag"
 	"fmt"
 	"net/http"
 
 	"github.com/AVENTER-UG/mesos-m3s/api"
 	"github.com/AVENTER-UG/mesos-m3s/mesos"
+	cfg "github.com/AVENTER-UG/mesos-m3s/types"
 	mesosutil "github.com/AVENTER-UG/mesos-util"
 	goredis "github.com/go-redis/redis/v8"
 
@@ -15,8 +19,11 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// MinVersion is the version number of this program
-var MinVersion string
+// BuildVersion of m3s
+var BuildVersion string
+
+// GitVersion is the revision and commit number
+var GitVersion string
 
 // init the redis cache
 func initCache() {
@@ -33,19 +40,41 @@ func initCache() {
 	config.RedisClient = client
 }
 
+// convert Base64 Encodes PEM Certificate to tls object
+func decodeBase64Cert(pemCert string) []byte {
+	sslPem, err := base64.URLEncoding.DecodeString(pemCert)
+	if err != nil {
+		logrus.Fatal("Error decoding SSL PEM from Base64: ", err.Error())
+	}
+	return sslPem
+}
+
 func main() {
+	// Prints out current version
+	var version bool
+	flag.BoolVar(&version, "v", false, "Prints current version")
+	flag.Parse()
+	if version {
+		fmt.Print(GitVersion)
+		return
+	}
+
 	util.SetLogging(config.LogLevel, config.EnableSyslog, config.AppName)
-	logrus.Println(config.AppName + " build " + MinVersion)
+	logrus.Println(config.AppName + " build " + BuildVersion + " git " + GitVersion)
 
 	listen := fmt.Sprintf(":%s", framework.FrameworkPort)
 
 	failoverTimeout := 5000.0
 	checkpoint := true
 	webuiurl := fmt.Sprintf("http://%s%s", framework.FrameworkHostname, listen)
+	if config.SSLCrt != "" && config.SSLKey != "" {
+		webuiurl = fmt.Sprintf("https://%s%s", framework.FrameworkHostname, listen)
+	}
 
 	framework.CommandChan = make(chan mesosutil.Command, 100)
 	config.Hostname = framework.FrameworkHostname
 	config.Listen = listen
+	config.Suppress = false
 
 	framework.State = map[string]mesosutil.State{}
 
@@ -68,24 +97,59 @@ func main() {
 	api.SetConfig(&config, &framework)
 
 	// load framework state from DB
-	key := api.GetRedisKey("framework")
+	key := api.GetRedisKey(framework.FrameworkName + ":framework")
 	if key != "" {
 		json.Unmarshal([]byte(key), &framework)
 	}
 
-	// load framework config from DB
-	key = api.GetRedisKey("framework_config")
+	// restore variable data from the old config
+	key = api.GetRedisKey(framework.FrameworkName + ":framework_config")
 	if key != "" {
-		json.Unmarshal([]byte(key), &config)
+		var oldconfig cfg.Config
+		json.Unmarshal([]byte(key), &oldconfig)
+		config.M3SBootstrapServerPort = oldconfig.M3SBootstrapServerPort
+		config.M3SBootstrapServerHostname = oldconfig.M3SBootstrapServerHostname
+		config.K3SServerPort = oldconfig.K3SServerPort
+		config.K3SServerURL = oldconfig.K3SServerURL
+		config.K3SAgentMax = oldconfig.K3SAgentMax
+		config.ETCDMax = oldconfig.ETCDMax
+
+		api.SaveConfig()
 	}
+
+	// set current m3s version
+	config.Version.M3SVersion.GitVersion = GitVersion
+	config.Version.M3SVersion.BuildDate = BuildVersion
 
 	// The Hostname should ever be set after reading the state file.
 	framework.FrameworkInfo.Hostname = &framework.FrameworkHostname
 
-	http.Handle("/", api.Commands())
+	server := &http.Server{
+		Addr:    listen,
+		Handler: api.Commands(),
+		TLSConfig: &tls.Config{
+			ClientAuth: tls.RequestClientCert,
+			MinVersion: tls.VersionTLS12,
+		},
+	}
+
+	if config.SSLCrt != "" && config.SSLKey != "" {
+		logrus.Debug("Enable TLS")
+		crt := decodeBase64Cert(config.SSLCrt)
+		key := decodeBase64Cert(config.SSLKey)
+		certs, err := tls.X509KeyPair(crt, key)
+		if err != nil {
+			logrus.Fatal("TLS Server Error: ", err.Error())
+		}
+		server.TLSConfig.Certificates = []tls.Certificate{certs}
+	}
 
 	go func() {
-		http.ListenAndServe(listen, nil)
+		if config.SSLCrt != "" && config.SSLKey != "" {
+			server.ListenAndServeTLS("", "")
+		} else {
+			server.ListenAndServe()
+		}
 	}()
 	logrus.Fatal(mesos.Subscribe())
 }