+++ bootstrap/bootstrap.sh
@@ -3,7 +3,6 @@
 cat /etc/resolv.conf
 
 apt-get update -y
-apt-get install -y jq containerd dnsmasq containernetworking-plugins tcpdump curl inetutils-ping iptables fuse-overlayfs procps bash iproute2 dnsutils net-tools systemctl
 mkdir -p /etc/cni/net.d
 
 export KUBERNETES_VERSION=v1.21.1
@@ -11,6 +10,7 @@ export INSTALL_K3S_VERSION=$KUBERNETES_VERSION+k3s1
 export INSTALL_K3S_SKIP_ENABLE=true
 export INSTALL_K3S_SKIP_START=true
 export KUBECONFIG=$MESOS_SANDBOX/kubeconfig.yaml
+export BRANCH=dev
 
 ## Export json as environment variables
 ## example: MESOS_SANDBOX_VAR='{ "CUSTOMER":"test-ltd" }'
@@ -19,17 +19,18 @@ for s in $(echo $MESOS_SANDBOX_VAR | jq -r "to_entries|map(\"\(.key)=\(.value|to
   export $s
 done
 
+## dockerd is a part of the uses avhost/ubuntu-m3s:focal docker image
+exec /usr/local/bin/dockerd &
 
-update-alternatives --set iptables /usr/sbin/iptables-legacy
-curl -sfL https://get.k3s.io | sh -
-curl https://raw.githubusercontent.com/AVENTER-UG/mesos-m3s/master/bootstrap/dashboard_auth.yaml > $MESOS_SANDBOX/dashboard_auth.yaml
-curl https://raw.githubusercontent.com/AVENTER-UG/mesos-m3s/master/bootstrap/dashboard_traefik.yaml > $MESOS_SANDBOX/dashboard_traefik.yaml
+curl -sfL https://get.k3s.io | INSTALL_K3S_VERSION=${INSTALL_K3S_VERSION} INSTALL_K3S_SKIP_ENABLE=${INSTALL_K3S_SKIP_ENABLE=$} INSTALL_K3S_SKIP_START=${INSTALL_K3S_SKIP_START} sh -s - --docker
+curl https://raw.githubusercontent.com/AVENTER-UG/mesos-m3s/${BRANCH}/bootstrap/dashboard_auth.yaml > $MESOS_SANDBOX/dashboard_auth.yaml
+curl https://raw.githubusercontent.com/AVENTER-UG/mesos-m3s/${BRANCH}/bootstrap/dashboard_traefik.yaml > $MESOS_SANDBOX/dashboard_traefik.yaml
 curl https://raw.githubusercontent.com/kubernetes/dashboard/v2.2.0/aio/deploy/recommended.yaml > $MESOS_SANDBOX/dashboard.yaml
 if [[ "$K3SFRAMEWORK_TYPE" == "server" ]]
 then
   curl -L http://dl.k8s.io/release/$KUBERNETES_VERSION/bin/linux/amd64/kubectl > $MESOS_SANDBOX/kubectl
-  curl https://raw.githubusercontent.com/AVENTER-UG/mesos-m3s/dev/bootstrap/server > $MESOS_SANDBOX/server
-  curl https://raw.githubusercontent.com/AVENTER-UG/mesos-m3s/dev/bootstrap/update.sh > $MESOS_SANDBOX/update
+  curl https://raw.githubusercontent.com/AVENTER-UG/mesos-m3s/${BRANCH}/bootstrap/server > $MESOS_SANDBOX/server
+  curl https://raw.githubusercontent.com/AVENTER-UG/mesos-m3s/${BRANCH}/bootstrap/update.sh > $MESOS_SANDBOX/update
   chmod +x $MESOS_SANDBOX/kubectl
   chmod +x $MESOS_SANDBOX/server
   chmod +x $MESOS_SANDBOX/update