+++ init.go
@@ -29,38 +29,44 @@ func init() {
 	framework.FrameworkInfoFilePath = util.Getenv("FRAMEWORK_STATEFILE_PATH", "/tmp")
 	framework.Username = os.Getenv("MESOS_USERNAME")
 	framework.Password = os.Getenv("MESOS_PASSWORD")
-	framework.MesosMasterServer = os.Getenv("MESOS_MASTER")
-	framework.MesosCNI = util.Getenv("MESOS_CNI", "weave")
+	framework.MesosMasterServer = util.Getenv("MESOS_MASTER", "127.0.0.1:5050")
+	framework.MesosCNI = os.Getenv("MESOS_CNI")
 	framework.PortRangeFrom, _ = strconv.Atoi(util.Getenv("PORTRANGE_FROM", "31000"))
 	framework.PortRangeTo, _ = strconv.Atoi(util.Getenv("PORTRANGE_TO", "32000"))
 	config.Principal = os.Getenv("MESOS_PRINCIPAL")
 	config.LogLevel = util.Getenv("LOGLEVEL", "info")
-	config.Domain = util.Getenv("DOMAIN", "local")
+	config.Domain = util.Getenv("DOMAIN", ".local")
 	config.K3SAgentMax, _ = strconv.Atoi(util.Getenv("K3S_AGENT_COUNT", "1"))
 	config.K3SServerMax, _ = strconv.Atoi(util.Getenv("K3S_SERVER_COUNT", "1"))
 	config.Credentials.Username = os.Getenv("AUTH_USERNAME")
 	config.Credentials.Password = os.Getenv("AUTH_PASSWORD")
 	config.AppName = "Mesos K3S Framework"
 	config.K3SCustomDomain = util.Getenv("K3S_CUSTOM_DOMAIN", "cloud.local")
-	config.K3SServerString = util.Getenv("K3S_SERVER_STRING", "/usr/local/bin/k3s server --cluster-cidr=10.2.0.0/16 --service-cidr=10.3.0.0/16 --cluster-dns=10.3.0.10 --snapshotter=native --flannel-backend=vxlan --flannel-iface=ethwe ")
-	config.K3SAgentString = util.Getenv("K3S_AGENT_STRING", "/usr/local/bin/k3s agent --snapshotter=native --flannel-iface=ethwe --flannel-backend=vxlan ")
-	config.ImageK3S = util.Getenv("IMAGE_K3S", "ubuntu:groovy")
-	config.ImageETCD = util.Getenv("IMAGE_ETCD", "docker.io/bitnami/etcd:latest")
+	config.K3SServerString = util.Getenv("K3S_SERVER_STRING", "/usr/local/bin/k3s server --cluster-cidr=10.2.0.0/16 --service-cidr=10.3.0.0/16 --cluster-dns=10.3.0.10  --kube-controller-manager-arg='leader-elect=false' --disable-cloud-controller --kube-scheduler-arg='leader-elect=false' --snapshotter=native --flannel-backend=vxlan ")
+	config.K3SAgentString = util.Getenv("K3S_AGENT_STRING", "/usr/local/bin/k3s agent --snapshotter=native --flannel-backend=vxlan ")
+	config.ImageK3S = util.Getenv("IMAGE_K3S", "avhost/ubuntu-m3s:focal")
+	config.ImageETCD = util.Getenv("IMAGE_ETCD", "docker.io/bitnami/etcd:3.5.1")
 	config.VolumeDriver = util.Getenv("VOLUME_DRIVER", "local")
 	config.VolumeK3SServer = util.Getenv("VOLUME_K3S_SERVER", "/data/k3s")
-	config.PrefixHostname = util.Getenv("PREFIX_HOSTNAME", "k3s")
 	config.K3SToken = util.Getenv("K3S_TOKEN", "123456789")
 	config.ETCDMax, _ = strconv.Atoi(util.Getenv("ETCD_COUNT", "1"))
-	config.BootstrapURL = util.Getenv("BOOTSTRAP_URL", "https://raw.githubusercontent.com/AVENTER-UG/go-mesos-framework-k3s/master/bootstrap/bootstrap.sh")
+	config.BootstrapURL = util.Getenv("BOOTSTRAP_URL", "https://raw.githubusercontent.com/AVENTER-UG/mesos-m3s/dev/bootstrap/bootstrap.sh")
 	config.DockerSock = os.Getenv("DOCKER_SOCK")
+	config.DockerSHMSize = util.Getenv("DOCKER_SHM_SIZE", "30gb")
 	config.M3SBootstrapServerPort, _ = strconv.Atoi(util.Getenv("M3S_BOOTSTRAP_SERVER_PORT", "6443"))
-	config.K3SCPU, _ = strconv.ParseFloat(util.Getenv("K3S_CPU", "0.1"), 64)
-	config.K3SMEM, _ = strconv.ParseFloat(util.Getenv("K3S_MEM", "1200"), 64)
+	config.K3SServerCPU, _ = strconv.ParseFloat(util.Getenv("K3S_SERVER_CPU", "0.1"), 64)
+	config.K3SServerMEM, _ = strconv.ParseFloat(util.Getenv("K3S_SERVER_MEM", "1200"), 64)
+	config.K3SAgentCPU, _ = strconv.ParseFloat(util.Getenv("K3S_AGENT_CPU", "0.1"), 64)
+	config.K3SAgentMEM, _ = strconv.ParseFloat(util.Getenv("K3S_AGENT_MEM", "1200"), 64)
 	config.ETCDCPU, _ = strconv.ParseFloat(util.Getenv("ETCD_CPU", "0.1"), 64)
 	config.ETCDMEM, _ = strconv.ParseFloat(util.Getenv("ETCD_MEM", "100"), 64)
+	config.ETCDDISK, _ = strconv.ParseFloat(util.Getenv("ETCD_DISK", "10000"), 64)
 	config.RedisServer = util.Getenv("REDIS_SERVER", "127.0.0.1:6379")
 	config.RedisPassword = os.Getenv("REDIS_PASSWORD")
 	config.RedisDB, _ = strconv.Atoi(util.Getenv("REDIS_DB", "1"))
+	config.SSLKey = os.Getenv("SSL_KEY_BASE64")
+	config.SSLCrt = os.Getenv("SSL_CRT_BASE64")
+	config.DockerCNI = util.Getenv("DOCKER_CNI", "bridge")
 
 	// if labels are set, unmarshel it into the Mesos Label format.
 	labels := os.Getenv("K3S_AGENT_LABELS")
@@ -94,6 +100,17 @@ func init() {
 		}
 	}
 
+	// if the constraint is set, determine which kind of
+	config.ETCDConstraint = util.Getenv("K3S_ETCD_CONSTRAINT", "")
+	if strings.Contains(config.ETCDConstraint, ":") {
+		constraint := strings.Split(config.ETCDConstraint, ":")
+
+		switch strings.ToLower(constraint[0]) {
+		case "hostname":
+			config.ETCDConstraintHostname = strings.ToLower(constraint[1])
+		}
+	}
+
 	// The comunication to the mesos server should be via ssl or not
 	if strings.Compare(os.Getenv("MESOS_SSL"), "true") == 0 {
 		framework.MesosSSL = true
@@ -101,4 +118,23 @@ func init() {
 		framework.MesosSSL = false
 	}
 
+	// Enable Docker Engine vor K3S instead critc
+	if strings.Compare(os.Getenv("K3S_DOCKER"), "true") == 0 {
+		config.K3SDocker = " --docker "
+	} else {
+		config.K3SDocker = ""
+	}
+
+	// Skip SSL Verification
+	if strings.Compare(os.Getenv("SKIP_SSL"), "true") == 0 {
+		config.SkipSSL = true
+	} else {
+		config.SkipSSL = false
+	}
+
+	// check if the domain starts with dot. if not, add one.
+	if !strings.HasPrefix(config.Domain, ".") {
+		tmp := config.Domain
+		config.Domain = "." + tmp
+	}
 }