+++ mesos/handle_update.go
@@ -13,13 +13,7 @@ import (
 
 // HandleUpdate will handle the offers event of mesos
 func HandleUpdate(event *mesosproto.Event) error {
-
 	logrus.Debug("HandleUpdate")
-	// unsuppress
-	revive := &mesosproto.Call{
-		Type: mesosproto.Call_REVIVE,
-	}
-	mesosutil.Call(revive)
 
 	update := event.Update