+++ types/types.go
@@ -12,7 +12,6 @@ import (
 type Config struct {
 	Principal                   string
 	LogLevel                    string
-	MinVersion                  string
 	AppName                     string
 	EnableSyslog                bool
 	Hostname                    string
@@ -36,17 +35,25 @@ type Config struct {
 	ImageETCD                   string
 	VolumeDriver                string
 	VolumeK3SServer             string
-	PrefixHostname              string
 	K3SToken                    string
 	ETCDMax                     int
 	DockerSock                  string
+	DockerSHMSize               string
 	BootstrapURL                string
 	M3SBootstrapServerHostname  string
 	M3SBootstrapServerPort      int
-	K3SCPU                      float64
-	K3SMEM                      float64
+	K3SServerCPU                float64
+	K3SServerMEM                float64
+	K3SServerDISK               float64
+	K3SAgentCPU                 float64
+	K3SAgentMEM                 float64
+	K3SAgentDISK                float64
+	K3SDocker                   string
 	ETCDCPU                     float64
 	ETCDMEM                     float64
+	ETCDDISK                    float64
+	ETCDConstraint              string
+	ETCDConstraintHostname      string
 	M3SStatus                   M3SStatus
 	MesosSandboxVar             string
 	RedisServer                 string
@@ -54,14 +61,20 @@ type Config struct {
 	RedisCTX                    context.Context
 	RedisPassword               string
 	RedisDB                     int
+	SkipSSL                     bool
+	SSLKey                      string
+	SSLCrt                      string
+	Version                     M3SVersion
+	Suppress                    bool
+	DockerCNI                   string
 }
 
 // M3SStatus store the current TaskState of the M3s services
 type M3SStatus struct {
-	Server []mesosproto.TaskState
-	Agent  []mesosproto.TaskState
+	Server map[string]string
+	Agent  map[string]string
 	API    string
-	Etcd   []mesosproto.TaskState
+	Etcd   map[string]string
 }
 
 // State will have the state of all tasks stated by this framework
@@ -82,16 +95,29 @@ type Count struct {
 	Running int // how many are running
 }
 
-// Version shows the current version of m3s
-type Version struct {
-	BootstrapBuild string `json:"bootstrap_build"`
-	M3sBuild       string `json:"m3s_build"`
-	M3sBersion     string `json:"m3s_version"`
-}
-
 // ErrorMsg hold the structure of error messages
 type ErrorMsg struct {
 	Message  string
 	Number   int
 	Function string
+}
+
+// M3SVersion hold the version numbers off the whole m3s stack
+type M3SVersion struct {
+	ClientVersion    versionInfo `json:"clientVersion"`
+	ServerVersion    versionInfo `json:"serverVersion"`
+	M3SVersion       versionInfo `json:"m3sVersion"`
+	BootstrapVersion versionInfo `json:"bootstrapVersion"`
+}
+
+type versionInfo struct {
+	Major        string `json:"major"`
+	Minor        string `json:"minor"`
+	GitVersion   string `json:"gitVersion"`
+	GitCommit    string `json:"gitCommit"`
+	GitTreeState string `json:"gitTreeState"`
+	BuildDate    string `json:"buildDate"`
+	GoVersion    string `json:"goVersion"`
+	Compiler     string `json:"compiler"`
+	Platform     string `json:"platform"`
 }